feat: add model struct

+++ domain/model/order_status_model.go
@@ -1,6 +1,6 @@
 package model
 
 type OrderStatus struct {
-	Id     int    `json:"id,omitempty"`
+	Id     uint64 `json:"id,omitempty"`
 	Status string `json:"status,omitempty"`
 }
